# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DP DB GATEWAY API
  version: "2.0"
  description: |
    数据库管理网关的 HTTP 接口。**每一条**已注册的路由都在本文件里 —— 由
    `TestOpenAPICoversEveryRoute` 静态守住:新增路由但没写进来,测试直接失败。

    ## 三套鉴权,不要混用

    | 面 | 前缀 | 鉴权 | 谁在调 |
    |---|---|---|---|
    | 控制台 | `/api/v1` | `Authorization: Bearer <JWT>` | 登录的人 |
    | 开放接口 | `/api/v1/open` | `Authorization: Bearer <key>.<secret>` | 外部系统 |
    | 回调 | `/api/v1/approvals/lark/callback` | 共享密钥(Bearer 或 `?secret=`) | 审批服务 |

    开放接口**不走**控制台的 IP 白名单:那个设置回答的是"哪些办公网可以打开控制台",
    把 CI runner 的出口段并进去等于为了一台机器放宽控制台的边界。每把凭据自带白名单。

    ## 响应包

    一律 HTTP 200,业务结果在 `code` 里 —— 包括错误。

    ```json
    { "code": 0, "msg": "ok", "data": { } }
    ```

    | code | 含义 | 客户端该做什么 |
    |---|---|---|
    | `0` | 成功 | 读 `data` |
    | `40001` | 参数错误 | `msg` 说明了哪里不对,直接展示 |
    | `40100` | 未登录 / Token 失效 / API 凭据无效 | 重新登录;开放接口检查 key.secret |
    | `40300` | 无权(菜单、能力矩阵、scope、非本人单据) | 展示 `msg`,不要重试 |
    | `40301` | 来源 IP 不在白名单 | 联系管理员放行 |
    | `42200` | 命令被拦截,已转审批 | `data.approvalNo` 是审批单号 |
    | `42600` | 脚本保存路径未配置 | 引导去【系统设置 · 网关】 |
    | `42601` | 导出保存路径未配置 | 同上 |
    | `42800` | 需要二次验证(TOTP) | 弹验证码输入,带 `mfaCode` 重试 |
    | `42900` | 登录尝试过于频繁 | 退避后再试 |
    | `50000` | 服务端错误 | 可重试;`msg` 不含内部细节 |

    ## 控制台的三层闸门

    每条控制台写接口要过:**菜单权限** → **能力矩阵(角色 × 能力 × 分层)** → **高危命令字典**。
    下面每个操作的 `summary` 都标了它要求的守卫,格式 `[菜单 · admin]`:

    - `[terminal]` 需要终端菜单;`[perms · admin]` 需要权限菜单**且**是平台管理员。
    - 没有标注 = 任何已登录用户可调。

...

...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 控制台会话 Token,由 `/auth/login` 签发。
    openApiKey:
      type: http
      scheme: bearer
      description: |
        开放接口凭据,格式 `<key>.<secret>`,在【系统设置 · 开放接口】签发。
        也接受 `X-Vela-Key` + `X-Vela-Secret` 两个请求头(给保留了 Authorization 的客户端)。
  schemas:
    R:
      type: object
      description: 统一响应包。
      required: [code, msg]
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string, example: ok }
        data: {}
    LoginReq:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: linwei@vela.io }
        password: { type: string, format: password }
        mfaCode: { type: string, description: 已绑定 TOTP 的账号在第二步提交 }
...
    RiskCheckReq:
      type: object
      required: [connectionId, sql]
      properties:
        connectionId: { type: integer, format: int64 }
        sql: { type: string }
    RiskCheckResp:
      type: object
      properties:
        risk: { type: string, enum: [high, mid, low] }
        action: { type: string, enum: [allow, approve, deny] }
        requiresApproval: { type: boolean }
        matchedRule: { type: string, description: 命中的规则;多条以 " + " 连接 }
        command: { type: string, description: 判定所依据的那条语句(批量时是最严的一条) }
    ExecReq:
      type: object
      required: [connectionId, sql]
      properties:
        connectionId: { type: integer, format: int64 }
        sql: { type: string, description: 可含多条语句;按最严的一条判定 }
        reason: { type: string, description: 转审批时写进工单 }
        mfaCode: { type: string, description: 目标分层要求二次验证时必填 }
        database: { type: string, description: 实例内的目标库;留空用连接默认库 }
    ExecResp:
      type: object
      properties:
        risk: { type: string }
        output: { type: string }
        rows: { type: integer }
        ms: { type: integer, description: 数据库执行耗时(实测,不含网关开销) }
        columns: { type: array, items: { type: string } }
        data: { type: array, items: { type: array, items: { type: string } } }
        truncated: { type: boolean }
        intercepted: { type: boolean }
        approvalNo: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: 登录(受 IP 白名单与登录频控约束)
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginReq' } } }
      responses:
        '200':
          description: OK / 40100 凭据错误 / 42800 需 MFA 验证码 / 42900 过于频繁
          content: { application/json: { schema: { $ref: '#/components/schemas/R' } } }
...
  /risk/check:
    post:
      tags: [terminal]
      summary: '[terminal] 三层预检,不执行'
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskCheckReq' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/RiskCheckResp' } } }
  /terminal/exec:
    post:
      tags: [terminal]
      summary: '[terminal] 执行命令(WS 不可用时的 REST 回退)'
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ExecReq' } } }
      responses:
        '200':
          description: OK / 42200 已转审批(返回 approvalNo)/ 42800 需二次验证 / 40300 拒绝
          content: { application/json: { schema: { $ref: '#/components/schemas/ExecResp' } } }
...
  /export/download:
    get:
      tags: [export]
      summary: '[terminal] 下载导出分卷'
      parameters: [{ name: file, in: query, required: true, schema: { type: string } }]
      responses: { '200': { description: 文件流 } }

  # ---------------------------------------------------------------- connections
...